add in Manga property List<Rating> and method getAverageRate(). Correct the mangaResponse add avgRate

+++ MangaST_1_11/src/main/java/com/example/mangast/rating/Rating.java
@@ -26,7 +26,7 @@ public class Rating {
     @GeneratedValue
     private Integer id;
 
-    private float rate;
+    private Double rate;
 
     @ManyToOne
     @JoinColumn(name = "user_id", nullable = false)
